+++ libpam/Makefile.am
@@ -18,8 +18,9 @@ include_HEADERS = include/security/_pam_compat.h \
 	include/security/pam_ext.h include/security/pam_modutil.h
 
 noinst_HEADERS = pam_prelude.h pam_private.h pam_tokens.h \
-		pam_modutil_private.h include/pam_cc_compat.h \
-		include/pam_inline.h include/test_assert.h
+		pam_modutil_private.h include/pam_assemble_line.h \
+		include/pam_cc_compat.h include/pam_inline.h \
+		include/test_assert.h
 
 libpam_la_LDFLAGS = -no-undefined -version-info 85:1:85
 libpam_la_LIBADD = @LIBAUDIT@ $(LIBPRELUDE_LIBS) $(ECONF_LIBS) @LIBDL@ @LTLIBINTL@
